[Timeseries] Refactor config.py to be able to import without pip install
